requirements.txt
aiohttp>=3.9.0
ib_insync>=0.9.86
discord.py>=2.3.0
python-dotenv>=1.0.0
#websocket-client>=1.6.0
pytz>=2023.3
hyperliquid-python-sdk>=0.21.0